+++ src/Command.ts
@@ -472,7 +472,7 @@ export default abstract class Command {
         return;
       }
 
-      const lowerCaseValue = value.toLowerCase();
+      const lowerCaseValue = value.toLowerCase().trim();
       if (lowerCaseValue === '@meid') {
         args.options[option] = accessToken.getUserIdFromAccessToken(token);
       }
@@ -576,9 +576,19 @@ export default abstract class Command {
       .replace(/([^\\])\\n/g, '$1\\\\\\n');
   }
 
-  public getCsvOutput(logStatement: any[]): string {
+  public getCsvOutput(logStatement: any[], options: GlobalOptions): string {
     const { stringify } = require('csv-stringify/sync');
     const cli = Cli.getInstance();
+
+    if (logStatement && logStatement.length > 0 && !options.query) {
+      logStatement.map(l => {
+        for (const x of Object.keys(l)) {
+          if (typeof l[x] === 'object') {
+            delete l[x];
+          }
+        }
+      });
+    }
 
     // https://csv.js.org/stringify/options/
     return stringify(logStatement, {
@@ -621,14 +631,16 @@ export default abstract class Command {
           `Property | Value`, os.EOL,
           `---------|-------`, os.EOL
         );
-        output.push(Object.keys(l).map(k => {
-          const value = l[k];
-          let stringValue = value;
-          if (typeof value === 'object') {
-            stringValue = JSON.stringify(value);
+        output.push(Object.keys(l).filter(x => {
+          if (!options.query && typeof l[x] === 'object') {
+            return;
           }
 
-          return `${md.escapeMd(k)} | ${md.escapeMd(stringValue)}`;
+          return x;
+        }).map(k => {
+          const value = l[k];
+
+          return `${md.escapeMd(k)} | ${md.escapeMd(value)}`;
         }).join(os.EOL), os.EOL);
         output.push(os.EOL);
       });