+++ src/m365/spfx/commands/project/project-upgrade/upgrade-1.17.0.ts
@@ -24,34 +24,42 @@ import { FN002020_DEVDEP_microsoft_rush_stack_compiler_4_5 } from './rules/FN002
 import { FN002022_DEVDEP_microsoft_eslint_plugin_spfx } from './rules/FN002022_DEVDEP_microsoft_eslint_plugin_spfx';
 import { FN002023_DEVDEP_microsoft_eslint_config_spfx } from './rules/FN002023_DEVDEP_microsoft_eslint_config_spfx';
 import { FN002024_DEVDEP_eslint } from './rules/FN002024_DEVDEP_eslint';
+import { FN007002_CFG_S_initialPage } from './rules/FN007002_CFG_S_initialPage';
 import { FN010001_YORC_version } from './rules/FN010001_YORC_version';
+import { FN010010_YORC_sdkVersions_teams_js } from './rules/FN010010_YORC_sdkVersions_teams_js';
+import { FN014008_CODE_launch_hostedWorkbench_type } from './rules/FN014008_CODE_launch_hostedWorkbench_type';
+import { FN014009_CODE_launch_hostedWorkbench_url } from './rules/FN014009_CODE_launch_hostedWorkbench_url';
 
 module.exports = [
-  new FN001001_DEP_microsoft_sp_core_library('1.17.0-beta.1'),
-  new FN001002_DEP_microsoft_sp_lodash_subset('1.17.0-beta.1'),
-  new FN001003_DEP_microsoft_sp_office_ui_fabric_core('1.17.0-beta.1'),
-  new FN001004_DEP_microsoft_sp_webpart_base('1.17.0-beta.1'),
-  new FN001011_DEP_microsoft_sp_dialog('1.17.0-beta.1'),
-  new FN001012_DEP_microsoft_sp_application_base('1.17.0-beta.1'),
-  new FN001014_DEP_microsoft_sp_listview_extensibility('1.17.0-beta.1'),
-  new FN001021_DEP_microsoft_sp_property_pane('1.17.0-beta.1'),
-  new FN001023_DEP_microsoft_sp_component_base('1.17.0-beta.1'),
-  new FN001024_DEP_microsoft_sp_diagnostics('1.17.0-beta.1'),
-  new FN001025_DEP_microsoft_sp_dynamic_data('1.17.0-beta.1'),
-  new FN001026_DEP_microsoft_sp_extension_base('1.17.0-beta.1'),
-  new FN001027_DEP_microsoft_sp_http('1.17.0-beta.1'),
-  new FN001028_DEP_microsoft_sp_list_subscription('1.17.0-beta.1'),
-  new FN001029_DEP_microsoft_sp_loader('1.17.0-beta.1'),
-  new FN001030_DEP_microsoft_sp_module_interfaces('1.17.0-beta.1'),
-  new FN001031_DEP_microsoft_sp_odata_types('1.17.0-beta.1'),
-  new FN001032_DEP_microsoft_sp_page_context('1.17.0-beta.1'),
-  new FN001013_DEP_microsoft_decorators('1.17.0-beta.1'),
-  new FN001034_DEP_microsoft_sp_adaptive_card_extension_base('1.17.0-beta.1'),
-  new FN002022_DEVDEP_microsoft_eslint_plugin_spfx('1.17.0-beta.1'),
-  new FN002023_DEVDEP_microsoft_eslint_config_spfx('1.17.0-beta.1'),
-  new FN002001_DEVDEP_microsoft_sp_build_web('1.17.0-beta.1'),
-  new FN002002_DEVDEP_microsoft_sp_module_interfaces('1.17.0-beta.1'),
-  new FN010001_YORC_version('1.17.0-beta.1'),
+  new FN001001_DEP_microsoft_sp_core_library('1.17.0'),
+  new FN001002_DEP_microsoft_sp_lodash_subset('1.17.0'),
+  new FN001003_DEP_microsoft_sp_office_ui_fabric_core('1.17.0'),
+  new FN001004_DEP_microsoft_sp_webpart_base('1.17.0'),
+  new FN001011_DEP_microsoft_sp_dialog('1.17.0'),
+  new FN001012_DEP_microsoft_sp_application_base('1.17.0'),
+  new FN001014_DEP_microsoft_sp_listview_extensibility('1.17.0'),
+  new FN001021_DEP_microsoft_sp_property_pane('1.17.0'),
+  new FN001023_DEP_microsoft_sp_component_base('1.17.0'),
+  new FN001024_DEP_microsoft_sp_diagnostics('1.17.0'),
+  new FN001025_DEP_microsoft_sp_dynamic_data('1.17.0'),
+  new FN001026_DEP_microsoft_sp_extension_base('1.17.0'),
+  new FN001027_DEP_microsoft_sp_http('1.17.0'),
+  new FN001028_DEP_microsoft_sp_list_subscription('1.17.0'),
+  new FN001029_DEP_microsoft_sp_loader('1.17.0'),
+  new FN001030_DEP_microsoft_sp_module_interfaces('1.17.0'),
+  new FN001031_DEP_microsoft_sp_odata_types('1.17.0'),
+  new FN001032_DEP_microsoft_sp_page_context('1.17.0'),
+  new FN001013_DEP_microsoft_decorators('1.17.0'),
+  new FN001034_DEP_microsoft_sp_adaptive_card_extension_base('1.17.0'),
+  new FN002022_DEVDEP_microsoft_eslint_plugin_spfx('1.17.0'),
+  new FN002023_DEVDEP_microsoft_eslint_config_spfx('1.17.0'),
+  new FN002001_DEVDEP_microsoft_sp_build_web('1.17.0'),
+  new FN002002_DEVDEP_microsoft_sp_module_interfaces('1.17.0'),
   new FN002020_DEVDEP_microsoft_rush_stack_compiler_4_5('0.4.0'),
-  new FN002024_DEVDEP_eslint('8.7.0')
+  new FN002024_DEVDEP_eslint('8.7.0'),
+  new FN007002_CFG_S_initialPage('https://{tenantDomain}/_layouts/workbench.aspx'),
+  new FN010001_YORC_version('1.17.0'),
+  new FN010010_YORC_sdkVersions_teams_js('2.9.1'),
+  new FN014008_CODE_launch_hostedWorkbench_type('edge'),
+  new FN014009_CODE_launch_hostedWorkbench_url('https://{tenantDomain}/_layouts/workbench.aspx')
 ];