+++ src/m365/spo/commands/file/file-move.ts
@@ -197,7 +197,7 @@ class SpoFileMoveCommand extends SpoCommand {
       await Cli.executeCommand(removeCommand as Command, { options: { ...removeOptions, _: [] } });
     }
     catch (err: any) {
-      if (err.error !== undefined && err.error.message !== undefined && err.error.message.includes('does not exist')) {
+      if (err !== undefined && err.message !== undefined && err.message.includes('does not exist')) {
 
       }
       else {